Add support for custom socket / stream dialers
This add custom dialers support for SocketsHttpHandler. This provide
support for client requiring to talk to an HTTP server listening on a
non-TCP socker (e.g.: a named pipe, Unix socket) and opens a way for
users to do end to end testing of their code base without having to open TCP ports.

Two flavors of dialers are supported:
- one returning both a Socket and a Stream (to allow socket level
manipulations such as Poll)
- one returning only a stream (in this case, socket level manipulations
are emulated by the existing HttpConnection class)